Create DNA Pride!
INPUT FORMAT

The first line of input contains T, the no of test cases. The next line of input contains N, the no of bases in each of the DNA sequence The line after that contains the DNA sequence.

OUTPUT FORMAT

For each test case output your answer on a new line.