Create Beer And Strings
First line of input contain Beer's favorite string X

Second line of the input contains a integer n number of string in the list. n line follows  line contains  string of the list.

Output

Output a single integer denoting the number of strings in the list which can be formed using letters of Beer's favorite string.