Create The Best Player@@
Input Format :

The first line contains N and T, the number of fans and the maximum number of fans Kalyani can meet. Each of the next N lines contains a string and an integer separated by a space. The string denotes the name of the fan while the integer depicts the fan quotient.

Output Format :

Output T lines, each containing the name of the fans selected. Fans with higher fan quotient should be outputted first and in case of a tie, the lexicographically minimum name should come first.

Constraints :




Name would only consist of characters in set [a-z]. It is not guaranteed that the names are distinct.

SAMPLE INPUT 
3 2
surbhi 3
surpanakha 3
shreya 5
SAMPLE OUTPUT 
shreya
surbhi
Explanation
We can see that the ordering would be {"shreya", "surbhi", "surpanakha"}. Hence the best two are shreya and surbhi.